feat - update the kb of an existing model

+++ backend/app/routes/knowledgeBase_routes.py
@@ -1,4 +1,5 @@
 from datetime import datetime
+from typing import Any
 import shutil
 from fastapi import APIRouter, BackgroundTasks, Depends, HTTPException
 from fastapi.responses import StreamingResponse
@@ -41,20 +42,19 @@ router = APIRouter(prefix="/knowledge_base")
 embedder = None
 
 def get_embedder():
-        global embedder
-        if embedder is None:
-            logging.info("Loading the Embedder")
-            os.makedirs(CACHE_DIR, exist_ok=True)
-            embedder = SentenceTransformer(
-                "sentence-transformers/paraphrase-multilingual-MiniLM-L12-v2",
-                cache_folder=CACHE_DIR
-            )
-            logging.info("Embedder loaded")
-        return embedder
-
+    global embedder
+    if embedder is None:
+        logging.info("Loading the Embedder")
+        os.makedirs(CACHE_DIR, exist_ok=True)
+        embedder = SentenceTransformer(
+            "sentence-transformers/paraphrase-multilingual-MiniLM-L12-v2",
+            cache_folder=CACHE_DIR
+        )
+        logging.info("Embedder loaded")
+    return embedder
 
 @router.get("/{llm_id}/status")
-def get_training_status(llm_id: int, db: Session = Depends(get_db)):
+def get_kbAttach_status(llm_id: int, db: Session = Depends(get_db)):
     """
     Get the status of a kb job.
     """
@@ -63,33 +63,31 @@ def get_training_status(llm_id: int, db: Session = Depends(get_db)):
         raise HTTPException(status_code=404, detail="KB job not found")
 
     status = kb_job.status
-    error_message = kb_job.error_message
     updated_at = kb_job.updated_at
 
     if status == "failed":
         llm = db.query(Llm).filter(Llm.id == llm_id).first()
         if llm:
-            if not error_message:
-                llm.error_message = "An unknown error occurred during kb creation."
-                db.delete(llm)
-        kb = db.query(KnowledgeBase).filter(KnowledgeBase.llm_id == llm_id).first()
-        if kb:
-            if kb.index_path:
-                if os.path.exists(kb.index_path):
-                    os.remove(kb.index_path)
-            db.delete(kb)
-        vector_store = db.query(VectorStore).filter(VectorStore.llm_id == llm_id).first()
-        if vector_store:
-            db.delete(vector_store)
+            if llm.kb_id:
+                vector_store = db.query(VectorStore).filter(VectorStore.kb_id == llm.kb_id).first()
+                if vector_store:
+                    db.delete(vector_store)
+                kb = db.query(KnowledgeBase).filter(KnowledgeBase.id == llm.kb_id).first()
+                if kb:
+                    if kb.index_path:
+                        if os.path.exists(kb.index_path):
+                            os.remove(kb.index_path)
+                    db.delete(kb)
+            db.delete(llm)
+            
         db.commit()
         db.refresh(kb_job)
 
     return {
         "status": status,
         "status_updated_at": updated_at,
-        "error_message": error_message if status == "failed" else None
+        "error_message": kb_job.error_message if status == "failed" else None
     }
-
 
 @router.post("/create", response_model=KnowledgeBaseResponse)
 def create_knowledge_base(
@@ -98,7 +96,7 @@ def create_knowledge_base(
     db: Session = Depends(get_db)
 ):
     """
-    Create a new knowledge base.
+    Attach a new knowledge base to an assistant.
     """
     logging.info(f"🚀 Received payload: {payload}")
     
@@ -150,6 +148,31 @@ def create_knowledge_base(
         logging.info(f"Created INDEXES_DIR at {INDEXES_DIR}")
 
     try:
+
+        if base_llm.is_attached_to_kb:
+            kb = db.query(KnowledgeBase).filter(KnowledgeBase.id == base_llm.kb_id).first()
+            if kb:
+                kb_job = KBJobModel(
+                    base_model_id=base_llm.id,
+                    new_model_id=base_llm.id,
+                    kb_id=kb.id,
+                    status="pending"
+                )
+                db.add(kb_job)
+                db.commit()
+                db.refresh(kb_job)      
+
+                background_tasks.add_task(update_kb_assistant_with_new_data, kb_job.id, kb.id, texts, db) 
+
+                return KnowledgeBaseResponse(
+                    msg="Assistant is being updated with the new data.",
+                    model_id=base_llm.id
+                )
+            
+            else:
+                logging.error(f"KnowledgeBase with ID {base_llm.kb_id} not found")
+                raise HTTPException(status_code=404, detail=f"LLM : {base_llm.name} seems to have a Knowledge Base attached, but the KB was not found.")
+        
         # Create new LLM
         new_llm = Llm(
             name=payload.modelName,
@@ -189,7 +212,7 @@ def create_knowledge_base(
         db.commit()
         logging.info(f"Knowledge Base job created with ID: {kb_job.id}")
 
-        background_tasks.add_task(init_kb_assistant, kb_job.id, new_llm.id, kb.id, vector_store.id, texts, db)
+        background_tasks.add_task(init_new_kb_assistant, kb_job.id, new_llm.id, kb.id, vector_store.id, texts, db)
 
         return KnowledgeBaseResponse(
             msg="Knowledge Base Assistant is being created.",
@@ -206,11 +229,158 @@ def create_knowledge_base(
     finally:
         db.close()
 
-def init_kb_assistant(kb_job_id: int, new_llm_id: int, kb_id: int, vector_store_id: int, texts: List[str], db: Session) -> Llm:
+def populate_vector_store(start_counter: int, vectors_data: dict, texts: List[str], index: Any) -> tuple[Any, dict]:
+    global embedder
+    if embedder is None:
+        embedder = get_embedder()
+
+    for text in texts:
+        if not text.strip():
+            continue
+        
+        logging.info(f"chunking text: {len(text)} characters")
+        chunks = chunk_by_tokens(text=text)
+        if not chunks:
+            logging.error(f"Error chunking text from {text}")
+            continue
+        logging.info(f"Chunks created: {len(chunks)} for text: {len(text)} characters")
+        
+        for chunk in chunks:
+            if not chunk.strip():
+                continue
+            logging.info(f"Encoding chunk: {chunk[:50]}...")
+            embeddings = embedder.encode(chunk, show_progress_bar=False, convert_to_tensor=True)
+            if embeddings is None or len(embeddings) == 0:
+                logging.error(f"Error encoding chunk: {chunk[:50]}")
+                continue
+            logging.info(f"Chunk encoded: {embeddings.shape}")
+            
+            # Store in vectors_data JSON and add to FAISS index
+            vectors_data[str(start_counter)] = chunk
+            try:
+                index.add_with_ids(embeddings.cpu().numpy().reshape(1, -1), [start_counter])
+            except Exception as e:
+                logging.error(f"Error adding vector to index: {e}")
+                continue
+            logging.info(f"Vector added to index with FAISS ID: {start_counter}")
+            start_counter += 1
+
+    del embedder
+    embedder = None
+    if torch.backends.mps.is_available():
+        torch.mps.empty_cache()
+    return (index, vectors_data)
+
+def update_kb_assistant_with_new_data(kb_job_id: int, kb_id: int, texts: List[str], db: Session) -> Llm:
+    try:
+        kb_job = db.query(KBJobModel).filter(KBJobModel.id == kb_job_id).first()
+        if not kb_job:
+            logging.error(f"KBJob with ID {kb_job_id} not found")
+            db.close()
+            raise HTTPException(status_code=404, detail="KBJob not found")
+        kb_job.status = "running"
+        kb_job.updated_at = datetime.now()
+        db.commit()
+    except Exception as e:
+        logging.error(f"Error fetching KBJob: {e}")
+        try:
+            db.close()
+        except Exception as e:
+            logging.error(f"Error rolling back DB session: {e}")
+            raise HTTPException(status_code=500, detail=f"Error closing DB session : {e}")
+        raise HTTPException(status_code=404, detail=f"Error manipulating DB session : {e}")
+
+    try:
+        llm = db.query(Llm).filter(Llm.id == kb_job.base_llm_id).first()
+        if not llm:
+            logging.error(f"LLM with ID {kb_job.base_llm_id} not found")
+            db.close()
+            raise HTTPException(status_code=404, detail="LLM not found")
+    except Exception as e:
+        logging.error(f"Error fetching LLM: {e}")
+        try:
+            db.close()
+        except Exception as e:
+            logging.error(f"Error rolling back DB session: {e}")
+            raise HTTPException(status_code=500, detail=f"Error closing DB session : {e}")
+        raise HTTPException(status_code=404, detail=f"Error manipulating DB session : {e}")
+
+    try:
+        kb = db.query(KnowledgeBase).filter(KnowledgeBase.id == kb_id).first()
+        if not kb:
+            logging.error(f"Knowledge Base with ID {kb_id} not found")
+            db.close()
+            raise HTTPException(status_code=404, detail="Knowledge Base not found")
+    except Exception as e:
+        logging.error(f"Error fetching Knowledge Base: {e}")
+        try:
+            db.close()
+        except Exception as e:
+            logging.error(f"Error rolling back DB session: {e}")
+            raise HTTPException(status_code=500, detail=f"Error closing DB session : {e}")
+        raise HTTPException(status_code=404, detail=f"Error manipulating DB session : {e}")
+
+    try:
+        vector_store = db.query(VectorStore).filter(VectorStore.kb_id == kb_id).first()
+        if not vector_store:
+            logging.error(f"VectorStore associated to KB {kb_id} not found")
+            db.close()
+            raise HTTPException(status_code=404, detail="VectorStore not found")
+    except Exception as e:
+        logging.error(f"Error fetching VectorStore: {e}")
+        try:
+            db.close()
+        except Exception as e:
+            logging.error(f"Error rolling back DB session: {e}")
+            raise HTTPException(status_code=500, detail=f"Error closing DB session : {e}")
+        raise HTTPException(status_code=404, detail=f"Error manipulating DB session : {e}")
+
+    initial_index = None
+    try:
+        initial_index = faiss.read_index(kb.index_path)
+    except:
+        logging.error(f"Error reading initial index")
+        raise HTTPException(status_code=404, detail=f"Index attached to LLM {llm.id} could not be read from disk")
+    
+    base_vectors_data = vector_store.vectors_data
+
+    try:
+        index = initial_index
+        start_index = max([int(k) for k in base_vectors_data.keys()]) + 1 if base_vectors_data else 0
+        try:
+            index, new_vectors_data = populate_vector_store(start_counter=start_index, vectors_data=base_vectors_data, texts=texts, index=index)
+        except:
+            raise
+        try:
+            faiss.write_index(index, kb.index_path)
+        except Exception as e2:
+            logging.error(f"Error writing index for KB {kb.id}: {e2}")
+            raise
+
+        vector_store.vectors_data = new_vectors_data
+        kb_job.status = "completed"
+        kb_job.updated_at = datetime.now()
+        db.commit()
+    except Exception as e:
+        logging.error(f"Error during KB update for LLM {llm.id}: {e}")
+        kb_job.status = "failed"
+        kb_job.updated_at = datetime.now()
+        vector_store.vectors_data = base_vectors_data
+        db.commit()
+
+        try:
+            faiss.write_index(initial_index, kb.index_path)
+        except Exception as e2:
+            logging.error(f"Error writing index for KB {kb.id}: {e2}")
+            raise HTTPException(status_code=404, detail=f"Error writing index for KB {kb.id}: {e2}")
+        
+    finally:
+        db.close()
+        
+def init_new_kb_assistant(kb_job_id: int, new_llm_id: int, kb_id: int, vector_store_id: int, texts: List[str], db: Session) -> Llm:
     """
     Initialize the knowledge base assistant by creating the FAISS index and storing vectors.
     """
-    global embedder
     try:
         kb_job = db.query(KBJobModel).filter(KBJobModel.id == kb_job_id).first()
         if not kb_job:
@@ -275,8 +445,6 @@ def init_kb_assistant(kb_job_id: int, new_llm_id: int, kb_id: int, vector_store_
         raise HTTPException(status_code=404, detail=f"Error manipulating DB session : {e}")
 
     try:
-        if embedder is None:
-            embedder = get_embedder()
         logging.info(f"Creating Index for Knowledge Base {kb.id} with {len(texts)} texts")
         try:
             index = faiss.IndexFlatL2(384)
@@ -288,38 +456,7 @@ def init_kb_assistant(kb_job_id: int, new_llm_id: int, kb_id: int, vector_store_
 
         # Prepare vectors data JSON for the VectorStore
         vectors_data = {}
-        faiss_id_counter = 0
-
-        for text in texts:
-            if not text.strip():
-                continue
-            
-            logging.info(f"chunking text: {len(text)} characters")
-            chunks = chunk_by_tokens(text=text)
-            if not chunks:
-                logging.error(f"Error chunking text from {text}")
-                continue
-            logging.info(f"Chunks created: {len(chunks)} for text: {len(text)} characters")
-            
-            for chunk in chunks:
-                if not chunk.strip():
-                    continue
-                logging.info(f"Encoding chunk: {chunk[:50]}...")
-                embeddings = embedder.encode(chunk, show_progress_bar=False, convert_to_tensor=True)
-                if embeddings is None or len(embeddings) == 0:
-                    logging.error(f"Error encoding chunk: {chunk[:50]}")
-                    continue
-                logging.info(f"Chunk encoded: {embeddings.shape}")
-                
-                # Store in vectors_data JSON and add to FAISS index
-                vectors_data[str(faiss_id_counter)] = chunk
-                try:
-                    index.add_with_ids(embeddings.cpu().numpy().reshape(1, -1), [faiss_id_counter])
-                except Exception as e:
-                    logging.error(f"Error adding vector to index: {e}")
-                    continue
-                logging.info(f"Vector added to index with FAISS ID: {faiss_id_counter}")
-                faiss_id_counter += 1
+        index, vectors_data = populate_vector_store(start_counter=0, vectors_data=vectors_data, texts=texts, index=index)
 
         vector_store.vectors_data = vectors_data
         db.commit()
@@ -383,7 +520,7 @@ def init_kb_assistant(kb_job_id: int, new_llm_id: int, kb_id: int, vector_store_
         kb_job.status = "completed"
         kb_job.updated_at = datetime.utcnow()
         db.commit()
-        logger.info(f"🎉 Knowledge Base {kb.id} created successfully with {faiss_id_counter} vectors!")
+        logger.info(f"🎉 Knowledge Base {kb.id} created successfully with {len(vectors_data)} vectors!")
         
         return new_llm
     
@@ -412,9 +549,4 @@ def init_kb_assistant(kb_job_id: int, new_llm_id: int, kb_id: int, vector_store_
             logging.error(f"Error updating KBJob status: {ex}")
 
     finally:
-        if embedder is not None:
-            del embedder
-            embedder = None
-            if torch.backends.mps.is_available():
-                torch.mps.empty_cache()
         db.close()
